Update job info route.
The route now takes a list of job IDs, as opposed to a single job ID, and returns a JSON array of the job infos.
This allows batching the job info requests for the front-end, as opposed to making a separate request for each job.
Additionally, the `status` field is now a dict that contains the progress of the job if available.